refactor: :fire: remove and reorganize the codebase
