+++ fixtures/json/pages.js
@@ -1,0 +1,42 @@
+export const pageSettings = {
+  forgotPassword: {
+    title: `Forgot Password?`,
+    description: `How could you forget your password?`,
+  },
+  login: {
+    title: `Login`,
+    description: `Access to your account`,
+  },
+  myAccount: {
+    title: `My Account`,
+    description: `Manage your account`,
+  },
+  resetPassword: {
+    title: `Reset Password`,
+    description: `Recover your account`,
+  },
+  blog: {
+    title: `From my Desk`,
+    description: `Archives concerning all matters web development and beyond`,
+  },
+  digitalGarden: {
+    title: `Digital Garden`,
+    description: `A collection of digital media`,
+  },
+  gallery: {
+    title: `Gallery`,
+    description: `A visual of all the things me!`,
+  },
+  models: {
+    title: `Models`,
+    description: `Airplanes, Tanks, Cars, its all here`,
+  },
+  home: {
+    title: `Home`,
+    description: `Programmer.  Model Enthusiast.  Intrigued Photographer.`,
+  },
+  parks: {
+    title: `Parks`,
+    description: `A collection of parks that I have visited`,
+  },
+};
