+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "dev": "next dev -p 6006",
-    "build": "SENTRY_LOG_LEVEL='debug' next build && next-sitemap",
+    "build": "ANALYZE=true SENTRY_LOG_LEVEL='debug' next build && next-sitemap",
     "postbuild": "next-sitemap",
     "lint": "next lint",
     "start": "NODE_OPTIONS='--max-old-space-size=1536' next start -p 6006",
@@ -17,8 +17,6 @@
   "author": "ChristopherLMiller",
   "license": "ISC",
   "dependencies": {
-    "@cloudinary/base": "^1.0.0-beta.4",
-    "@cloudinary/react": "^1.3.0",
     "@hcaptcha/react-hcaptcha": "^1.3.0",
     "@mapbox/rehype-prism": "^0.8.0",
     "@mavrin/remark-typograf": "^2.2.0",
@@ -26,28 +24,23 @@
     "@mdx-js/react": "^2.1.1",
     "@mdx-js/runtime": "^1.6.22",
     "@next/mdx": "^12.1.6",
-    "@sentry/integrations": "^6.19.7",
     "@sentry/nextjs": "^6.19.7",
     "@types/leaflet": "^1.7.10",
     "assert": "^2.0.0",
     "date-fns": "^2.28.0",
     "dotenv-flow": "^3.2.0",
-    "exif": "^0.6.0",
-    "exif-js": "^2.3.0",
     "formik": "^2.2.9",
-    "framer-motion": "^6.3.3",
-    "graphql": "^16.4.0",
+    "framer-motion": "^6.3.4",
+    "graphql": "^16.5.0",
     "graphql-request": "^4.2.0",
     "graphql-tag": "^2.12.6",
     "leaflet": "^1.8.0",
     "next": "^12.1.6",
     "next-auth": "^4.3.4",
-    "next-offline": "^5.0.5",
     "next-pwa": "^5.5.2",
     "next-seo": "^5.4.0",
-    "next-sitemap": "^2.5.20",
+    "next-sitemap": "^2.5.25",
     "next-transpile-modules": "^9.0.0",
-    "node-fetch": "^3.2.4",
     "normalize.css": "^8.0.1",
     "nprogress": "^0.2.0",
     "path-browserify": "^1.0.1",
@@ -59,13 +52,13 @@
     "react-icons": "^4.3.1",
     "react-leaflet": "^4.0.0",
     "react-modal": "^3.15.1",
-    "react-query": "^3.38.1",
+    "react-query": "^3.39.0",
     "react-responsive-masonry": "^2.1.4",
     "react-share": "^4.4.0",
     "react-snowfall": "^1.1.2",
-    "react-spinners": "^0.11.0",
+    "react-spinners": "^0.12.0",
     "react-toast-notifications": "^2.5.1",
-    "react-youtube": "^9.0.1",
+    "react-youtube": "^9.0.2",
     "rehype-react": "^7.1.1",
     "remark": "^14.0.2",
     "remark-emoji": "^3.0.2",
@@ -80,44 +73,40 @@
     "remark-rehype": "^10.1.0",
     "remark-sub-super": "^1.0.20",
     "remark-unwrap-images": "^3.0.1",
-    "sharp": "^0.30.4",
     "simple-react-lightbox": "^3.6.9-0",
     "styled-components": "^5.3.5",
-    "unified": "^10.1.2",
     "yup": "^0.32.11"
   },
   "devDependencies": {
-    "@babel/core": "7.17.10",
-    "@babel/eslint-parser": "7.17.0",
+    "@babel/core": "7.18.2",
+    "@babel/eslint-parser": "7.18.2",
     "@graphql-codegen/cli": "2.6.2",
-    "@graphql-codegen/near-operation-file-preset": "2.2.9",
-    "@graphql-codegen/typescript": "2.4.8",
-    "@graphql-codegen/typescript-operations": "2.3.5",
-    "@graphql-codegen/typescript-react-apollo": "3.2.11",
-    "@graphql-codegen/typescript-react-query": "3.5.9",
+    "@graphql-codegen/near-operation-file-preset": "2.2.12",
+    "@graphql-codegen/typescript": "2.4.11",
+    "@graphql-codegen/typescript-operations": "2.4.0",
+    "@graphql-codegen/typescript-react-query": "3.5.12",
+    "@next/bundle-analyzer": "^12.1.6",
     "@trivago/prettier-plugin-sort-imports": "3.2.0",
-    "@types/gtag.js": "0.0.10",
     "@types/hcaptcha__react-hcaptcha": "0.1.5",
-    "@types/node": "17.0.31",
-    "@types/react": "18.0.8",
-    "@typescript-eslint/eslint-plugin": "5.22.0",
-    "@typescript-eslint/parser": "5.22.0",
+    "@types/node": "17.0.35",
+    "@types/react": "18.0.9",
+    "@typescript-eslint/eslint-plugin": "5.26.0",
+    "@typescript-eslint/parser": "5.26.0",
     "airbnb": "0.0.2",
     "babel-loader": "8.2.5",
     "babel-plugin-styled-components": "2.0.7",
     "cross-env": "7.0.3",
-    "eslint": "8.14.0",
+    "eslint": "8.16.0",
     "eslint-config-airbnb": "19.0.4",
     "eslint-config-next": "12.1.6",
     "eslint-config-prettier": "8.5.0",
     "eslint-plugin-import": "2.26.0",
     "eslint-plugin-jsx-a11y": "6.5.1",
     "eslint-plugin-prettier": "4.0.0",
-    "eslint-plugin-react": "7.29.4",
+    "eslint-plugin-react": "7.30.0",
     "eslint-plugin-react-hooks": "4.5.0",
     "prettier": "^2.6.2",
-    "react-is": "18.1.0",
-    "typescript": "4.6.4"
+    "typescript": "4.7.2"
   },
   "nextBundleAnalysis": {
     "budget": 358400,
