+++ pages/account/my-account.tsx
@@ -1,5 +1,6 @@
 import { Grid, GridItem } from "@components/Grid";
-import { defaultImage, pageSettings } from "config";
+import { pageSettings } from "@fixtures/json/pages";
+import { defaultImage } from "config";
 import { GetServerSideProps, NextPage } from "next";
 import { getSession, useSession } from "next-auth/react";
 import { NextSeo } from "next-seo";
