+++ pages/gallery/album/[slug].tsx
@@ -4,7 +4,8 @@ import CloudinaryImage from "@components/Images/CloudinaryImage";
 import Markdown from "@components/Markdown";
 import ShareButtons from "@components/ShareButtons";
 import Table from "@components/Table";
-import { lightboxOptions, pageSettings } from "config";
+import { pageSettings } from "@fixtures/json/pages";
+import { lightboxOptions } from "config";
 import { formatRelative } from "date-fns";
 import { GetServerSideProps, NextPage } from "next";
 import { getSession, useSession } from "next-auth/react";
