+++ pages/models/index.tsx
@@ -2,7 +2,8 @@ import { Grid } from "@components/Grid";
 import Loader from "@components/Loader";
 import ModelCard from "@components/ModelCard";
 import Paginator from "@components/Paginator";
-import { defaultImage, pageSettings } from "config";
+import { pageSettings } from "@fixtures/json/pages";
+import { defaultImage } from "config";
 import { NextPage } from "next";
 import { useSession } from "next-auth/react";
 import { NextSeo } from "next-seo";
