+++ pages/parks.tsx
@@ -1,4 +1,5 @@
-import { ApiEndpoint, defaultImage, pageSettings } from "config";
+import { pageSettings } from "@fixtures/json/pages";
+import { ApiEndpoint, defaultImage } from "config";
 import { NextPage } from "next";
 import { NextSeo } from "next-seo";
 import dynamic from "next/dynamic";
