+++ src/components/Panel/styles.tsx
@@ -21,6 +21,7 @@ export const StyledPanel = styled(motion.div)`
     z-index: -1;
   }
 
+  color: var(--color-black-80);
   height: min-content;
   padding: ${(props) => props.padding};
   position: relative;
@@ -30,5 +31,59 @@ export const StyledPanel = styled(motion.div)`
 
   img {
     max-width: 100%;
+    display: inline-block;
+  }
+
+  p {
+    break-inside: avoid;
+    padding-block-end: 10px;
+    padding-block-start: 10px;
+    text-indent: 2em;
+  }
+
+  h3,
+  h4,
+  h5,
+  h6 {
+    margin-block: 1rem;
+    position: relative;
+    padding-inline-start: 2rem;
+    color: var(--color-black-60);
+    font-family: var(--font-block);
+    text-transform: uppercase;
+    font-style: italic;
+
+    :before {
+      content: "\\A";
+      position: absolute;
+      left: 5px;
+      width: 50%;
+      height: 100%;
+      border-inline-start: 3px solid var(--color-gold-transparent);
+      border-block-end: 5px solid var(--color-red-dark);
+      opacity: 0.7;
+      transform: skewX(-12deg);
+    }
+  }
+
+  a {
+    color: var(--color-red-intermediate);
+    box-shadow: var(--box-shadow-inset-0);
+
+    :hover {
+      color: var(--color-gold-transparent);
+      box-shadow: var(--box-shadow-inset-2);
+    }
+    img {
+      width: 33%;
+      max-width: 33%;
+      display: inline-block;
+    }
+  }
+
+  blockquote {
+    border-inline-start: 10px solid var(--color-grey-intermediate);
+    padding-inline-start: 1rem;
+    background: var(--color-black-20);
   }
 `;
