+++ src/components/index.tsx
@@ -1,6 +1,7 @@
 export * from "./Card";
 export * from "./Caret";
 export * from "./Grid";
+export * from "./Loader";
 export * from "./ModelCard";
 export * from "./Paginator";
 export * from "./Panel";
