+++ src/utils/index.ts
@@ -1,13 +1,8 @@
 export * from "./auth";
 export * from "./countWords";
 export * from "./filterNavigation";
-export * from "./getBuildTime";
-export * from "./getExcerpt";
 export * from "./getYoutubeVideoId";
-export * from "./isClientServer";
 export * from "./isDev";
 export * from "./makeDurationFriendly";
-export * from "./redirectIfProd";
 export * from "./sendEmail";
 export * from "./timeToRead";
-export * from "./truncate";
