--- src/utils/isClientServer.ts
@@ -1,3 +1,0 @@
-export const isClient = (): boolean => window !== undefined;
-
-export const isServer = (): boolean => window === undefined;
